make chess the index route
